Fix bug

+++ application/src/main/java/common/sink/MeasureSink.java
@@ -212,11 +212,11 @@ public class MeasureSink extends BaseSink {
                 sb.append(latency.getPercentile(99) + "\n");
                 w.write(sb.toString());
                 w.write("Percentile\t Latency\n");
-                w.write(String.format("%d\t" +
+                w.write(String.format("%f\t" +
                                 "%-10.4f\t"
                         , 0.5,latency.getPercentile(0.5)));
-                for (int i = 20; i < 100; i += 20){
-                    String output = String.format("%d\t" +
+                for (double i = 20; i < 100; i += 20){
+                    String output = String.format("%f\t" +
                                     "%-10.4f\t"
                             , i,latency.getPercentile(i));
                     w.write(output + "\n");
